+++ Engine/src/Platform/Vulkan/Internal/VulkanSwapchain.cpp
@@ -1,0 +1,171 @@
+#include "Platform/Vulkan/Internal/VulkanSwapchain.h"
+
+#include <algorithm>
+
+#include "Utils/Log.h"
+#include "Platform/Vulkan/Internal/VulkanUtils.h"
+
+namespace aero3d {
+
+static SwapChainSupportDetails QuerySwapChainSupport(VkPhysicalDevice device, VkSurfaceKHR surface)
+{
+    SwapChainSupportDetails details;
+
+    vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, surface, &details.Capabilities);
+
+    uint32_t formatCount;
+    vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, nullptr);
+    if (formatCount != 0)
+    {
+        details.Formats.resize(formatCount);
+        vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, details.Formats.data());
+    }
+
+    uint32_t presentModeCount;
+    vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, nullptr);
+    if (presentModeCount != 0)
+    {
+        details.PresentModes.resize(presentModeCount);
+        vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, details.PresentModes.data());
+    }
+
+    return details;
+}
+
+static VkSurfaceFormatKHR ChooseSwapSurfaceFormat(const std::vector<VkSurfaceFormatKHR>& availableFormats)
+{
+    for (const auto& format : availableFormats)
+    {
+        if (format.format == VK_FORMAT_B8G8R8A8_SRGB &&
+            format.colorSpace == VK_COLOR_SPACE_SRGB_NONLINEAR_KHR)
+        {
+            return format;
+        }
+    }
+    return availableFormats[0];
+}
+
+static VkPresentModeKHR ChooseSwapPresentMode(const std::vector<VkPresentModeKHR>& availablePresentModes)
+{
+    for (const auto& mode : availablePresentModes)
+    {
+        if (mode == VK_PRESENT_MODE_MAILBOX_KHR)
+        {
+            return mode;
+        }
+    }
+    return VK_PRESENT_MODE_FIFO_KHR;
+}
+
+static VkExtent2D ChooseSwapExtent(const VkSurfaceCapabilitiesKHR& capabilities, SDL_Window* window)
+{
+    if (capabilities.currentExtent.width != UINT32_MAX)
+    {
+        return capabilities.currentExtent;
+    }
+    else
+    {
+        int width, height;
+        SDL_GetWindowSizeInPixels(window, &width, &height);
+
+        VkExtent2D actualExtent = {
+            static_cast<uint32_t>(width),
+            static_cast<uint32_t>(height)
+        };
+
+        actualExtent.width = std::clamp(actualExtent.width,
+            capabilities.minImageExtent.width,
+            capabilities.maxImageExtent.width);
+
+        actualExtent.height = std::clamp(actualExtent.height,
+            capabilities.minImageExtent.height,
+            capabilities.maxImageExtent.height);
+
+        return actualExtent;
+    }
+}
+
+VulkanSwapchain::VulkanSwapchain()
+    : m_Swapchain(VK_NULL_HANDLE), m_Surface(VK_NULL_HANDLE), m_Window(nullptr),
+    m_Device(VK_NULL_HANDLE), m_Extent(), m_ImageFormat(VK_FORMAT_UNDEFINED)
+{
+}
+
+VulkanSwapchain::~VulkanSwapchain()
+{
+}
+
+bool VulkanSwapchain::Init(const VulkanPhysicalDevice& physDevice, VkSurfaceKHR surface,
+                            SDL_Window* window, VkDevice device)
+{
+    m_Device = device;
+    m_Surface = surface;
+    m_Window = window;
+
+    CreateSwapchain(physDevice);
+
+    return true;
+}
+
+void VulkanSwapchain::Shutdown()
+{
+    vkDestroySwapchainKHR(m_Device, m_Swapchain, nullptr);
+}
+
+void VulkanSwapchain::CreateSwapchain(const VulkanPhysicalDevice& physDevice)
+{
+    SwapChainSupportDetails swapChainSupport = QuerySwapChainSupport(physDevice.Device, m_Surface);
+
+    VkSurfaceFormatKHR surfaceFormat = ChooseSwapSurfaceFormat(swapChainSupport.Formats);
+    VkPresentModeKHR presentMode = ChooseSwapPresentMode(swapChainSupport.PresentModes);
+    VkExtent2D extent = ChooseSwapExtent(swapChainSupport.Capabilities, m_Window);
+
+    uint32_t imageCount = swapChainSupport.Capabilities.minImageCount + 1;
+    if (swapChainSupport.Capabilities.maxImageCount > 0 &&
+        imageCount > swapChainSupport.Capabilities.maxImageCount)
+    {
+        imageCount = swapChainSupport.Capabilities.maxImageCount;
+    }
+
+    VkSwapchainCreateInfoKHR createInfo{};
+    createInfo.sType = VK_STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR;
+    createInfo.surface = m_Surface;
+
+    createInfo.minImageCount = imageCount;
+    createInfo.imageFormat = surfaceFormat.format;
+    createInfo.imageColorSpace = surfaceFormat.colorSpace;
+    createInfo.imageExtent = extent;
+    createInfo.imageArrayLayers = 1;
+    createInfo.imageUsage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
+
+    DeviceQueueFamilyIndices indices = physDevice.QueueFamilyIndices;
+    uint32_t queueFamilyIndices[] = { indices.GraphicsFamily.value(), indices.PresentFamily.value() };
+
+    if (indices.GraphicsFamily != indices.PresentFamily)
+    {
+        createInfo.imageSharingMode = VK_SHARING_MODE_CONCURRENT;
+        createInfo.queueFamilyIndexCount = 2;
+        createInfo.pQueueFamilyIndices = queueFamilyIndices;
+    }
+    else
+    {
+        createInfo.imageSharingMode = VK_SHARING_MODE_EXCLUSIVE;
+    }
+
+    createInfo.preTransform = swapChainSupport.Capabilities.currentTransform;
+    createInfo.compositeAlpha = VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR;
+    createInfo.presentMode = presentMode;
+    createInfo.clipped = VK_TRUE;
+    createInfo.oldSwapchain = VK_NULL_HANDLE;
+
+    A3D_CHECK_VKRESULT(vkCreateSwapchainKHR(m_Device, &createInfo, nullptr, &m_Swapchain));
+
+    vkGetSwapchainImagesKHR(m_Device, m_Swapchain, &imageCount, nullptr);
+    m_Images.resize(imageCount);
+    vkGetSwapchainImagesKHR(m_Device, m_Swapchain, &imageCount, m_Images.data());
+
+    m_ImageFormat = surfaceFormat.format;
+    m_Extent = extent;
+}
+
+} // namespace aero3d
